Changed backup/restore to use use rsync
It made sense to use Unison for real-time synchronization when we were mirroring to a VirtualBox shared folder. However due to problems with the shared folders we moved to backing up to a second virtual disk instead.

Given we no longer need real-time synchronisation it makes sense to switch to rsync as it's more robust and better supported.